code/api/requirements.txt
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
SQLAlchemy==1.4.27
Werkzeug==2.0.2
